sizeof

+++ 0x00-hello_world/6-size.c
@@ -11,11 +11,11 @@ printf("Size of a char: %i", sizeof(char));
 printf(" byte(s)\n");
 printf("Size of an int: %i", sizeof(int));
 printf(" byte(s)\n");
-printf("Size of a long int: %li", sizeof(long));
+printf("Size of a long int: %d", sizeof(long));
 printf(" byte(s)\n");
-printf("Size of a long long int: %li", sizeof(long));
+printf("Size of a long long int: %d", sizeof(double));
 printf(" byte(s)\n");
-printf("Size of a float: %i", sizeof(float));
+printf("Size of a float: %d", sizeof(float));
 printf(" byte(s)\n");
 
 return (0);
